AND:fix: modify logic to start hiking (start hiking button click -> modals ->...

See merge request s08-final/S08P31A102!466

+++ android/hikingdom/app/src/main/java/com/example/hikingdom/data/remote/hiking/HikingData.kt
@@ -19,9 +19,9 @@ data class MountainResponse(
 )
 
 data class Meetup(
-    @SerializedName("meetupId") val meetupId: Int,
+    @SerializedName("meetupId") val meetupId: Long,
     @SerializedName("meetupName") val meetupName: String,
-    @SerializedName("mountainId") val mountainId: Int,
+    @SerializedName("mountainId") val mountainId: Long,
     @SerializedName("mountainName") val mountainName: String,
     @SerializedName("mountainSummitLat") val mountainSummitLat: Double,
     @SerializedName("mountainSummitLng") val mountainSummitLng: Double,
